Раскладывает на модули
